refactor: relocate filters into outputs/, make theme private
Two surgical root tidies (Codex-reviewed):
- filters.py -> outputs/filters.py: it is output-side post-processing of
  prediction/control vectors, not a core primitive. outputs/ already owns
  'vectors leaving the app'. Re-exported from myogestic.outputs (OneEuroFilter,
  GaussianFilter, IdentityFilter, VectorFilter, make_filter). Callers + docs
  autodoc/xrefs repointed to myogestic.outputs.filters.
- theme.py -> _theme.py: only core.py imports it, not re-exported, runtime
  styling plumbing -> private per convention.

Root now holds only the public spine (core, stream, contracts, grid,
edge_trigger) plus private runtime support (_browser, _platform, _theme).
