+++ docs/api/filters.md
@@ -1,19 +1,19 @@
 # Filters
 
-Output-side smoothing filters for prediction-output control vectors. Custom filters implement the [`VectorFilter`][myogestic.filters.VectorFilter] protocol below. See [Post-process predictions](../how-to/post-process-output.md) for tuning guidance.
+Output-side smoothing filters for prediction-output control vectors. Custom filters implement the [`VectorFilter`][myogestic.outputs.filters.VectorFilter] protocol below. See [Post-process predictions](../how-to/post-process-output.md) for tuning guidance.
 
 ## The protocol
 
-::: myogestic.filters.VectorFilter
+::: myogestic.outputs.filters.VectorFilter
 
 ## Built-in filters
 
-::: myogestic.filters.OneEuroFilter
+::: myogestic.outputs.filters.OneEuroFilter
 
-::: myogestic.filters.GaussianFilter
+::: myogestic.outputs.filters.GaussianFilter
 
-::: myogestic.filters.IdentityFilter
+::: myogestic.outputs.filters.IdentityFilter
 
 ## Factory
 
-::: myogestic.filters.make_filter
+::: myogestic.outputs.filters.make_filter
