+++ docs/how-to/index.md
@@ -14,7 +14,7 @@ Recipes for specific tasks. Each guide assumes you've worked through [Getting St
 - [Record and replay](record-and-replay.md) - capture sessions, read them back programmatically.
 - [Record good training data](record-good-training-data.md) - cycle-style recording, how many cycles you actually need, verifying templates before training.
 - [Feature extraction cookbook](feature-extraction-cookbook.md) - copy-paste `@pipeline.extract` snippets (RMS+MAV, bandpass+envelope, spectral, sliding RMS, onset detection, multi-stream fusion).
-- [Post-process predictions](post-process-output.md) - `FilterControl` and `myogestic.filters` for output smoothing.
+- [Post-process predictions](post-process-output.md) - `FilterControl` and `myogestic.outputs.filters` for output smoothing.
 - [Integrate the Virtual Hand](integrate-vhi.md) - `myogestic.vhi.interfaces.virtual_hand` and the launcher pattern.
 
 ## Operations
